Project 4
Had trouble testing some of it.  Never seemed to download items from
other people but that should not be an issue with my code.   I just
think I was running it wrong.  I did not implement a lot for defending
attacks and I implemented 2 ways to attack other peers.   The only thing
that could really be done is to try to figure out how to test it.  I
know a lot of people had problems testing it from online. Also it has to
be run on seas and apparently lnx srv 1.   I also had to go in and
change the server addresses like it said on piazza.  I may have
potentially done that incorreclty which is why I am struggling on
testing it.